refactor: break out Tx structs into their own files
fix: missed moving ContractCreateTx to tx_contracts.go when splitting up transactions.go
